# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Parking Space Sharing API
  version: unstable
  description: API specification for the parking space sharing application.
servers:
  - url: https://xpark-staging.ersei.net/api/unstable
    description: Staging Server
...
paths:
...
  /parking-spaces:
    post:
      tags:
        - Parking Spaces
      summary: Create a new parking space
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/ParkingSpaceCreateRequest.yaml'
      responses:
        '201':
          description: Parking space created
          content:
            application/json:
              schema:
                $ref: './schemas/ParkingSpace.yaml'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  /parking-spaces/{id}:
    get:
      tags:
        - Parking Spaces
      summary: Get parking space details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Parking space details
          content:
            application/json:
              schema:
                $ref: './schemas/ParkingSpace.yaml'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
    patch:
      tags:
        - Parking Spaces
      summary: Update parking space
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/ParkingSpaceUpdateRequest.yaml'
      responses:
        '200':
          description: Parking space updated
          content:
            application/json:
              schema:
                $ref: './schemas/ParkingSpace.yaml'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Parking space not found
    delete:
      tags:
        - Parking Spaces
      summary: Delete parking space
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Parking space deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Parking space not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer